Cleanup and fixes
Simplified and better explained a piece of code early on in the MMU3 response detection responsible for updating the error during the ATTENTION state
Deduplication for the lines that trigger ATTENTION on the MMU3. Some errors send both lines, or spam one
Better explanation of some MMU3 lines in Mmu.py
Minor formatting edits

Known Issue: If events sent to the event bus are delayed, the lines that check the current state may be checking an old state, breaking checks. Currently waiting on feedback for a fix for this.